requirements.txt
fastapi
uvicorn
boto3
openai
python-dotenv
pydantic